+++ src/specarray/specarray.py
@@ -14,10 +14,6 @@ from spectral.io.bilfile import BilFile
 from .io.specim_folder import from_specim_folder
 
 
-def trapz(x, y):
-    return np.trapz(y, x)
-
-
 @dataclass
 class SpecArray:
     """Class for Specim hyperspectral Camera data"""
@@ -34,7 +30,15 @@ class SpecArray:
     @classmethod
     def from_folder(cls, folder: Path):
         """Create a SpecArray from a folder"""
-        return cls(*from_specim_folder(folder))
+        capture, metadata, wavelengths, black, white, capture_spectral = from_specim_folder(folder)
+        return cls(
+            capture=capture,
+            metadata=metadata,
+            wavelengths=wavelengths,
+            black=black,
+            white=white,
+            _capture_spectral=capture_spectral,
+        )
 
     def __len__(self):
         """Return the number of records"""
@@ -62,9 +66,9 @@ class SpecArray:
     def spectral_albedo(self) -> xr.DataArray:
         """Calculate and return the spectral albedo. The values are limited to 0 and 1."""
         if self.has_black and self.has_white:
-            spectral_albedo = (self.capture - self.black.mean(dim="sample")) / (
-                self.white.mean(dim="sample") - self.black.mean(dim="sample")
-            )
+            black_mean = self.black.mean(dim="sample")
+            white_mean = self.white.mean(dim="sample")
+            spectral_albedo = (self.capture - black_mean) / (white_mean - black_mean)
             spectral_albedo = xr.where(spectral_albedo < 0.0, 0.0, spectral_albedo)
             spectral_albedo = xr.where(spectral_albedo > 1.0, 1.0, spectral_albedo)
             spectral_albedo.name = "spectral albedo"
@@ -76,20 +80,17 @@ class SpecArray:
     def broadband_albedo(self) -> xr.DataArray:
         """Calculate and return the broadband albedo"""
         if self.has_black and self.has_white:
-            spectral_albedo = self.spectral_albedo.chunk({"sample": -1})
-            broadband_albedo = xr.apply_ufunc(
-                trapz,
-                spectral_albedo.coords["wavelength"],
-                spectral_albedo.transpose(
+            broadband_albedo = np.trapz(
+                self.spectral_albedo.transpose(
                     "sample",
                     "point",
                     "wavelength",
                 ),
-                input_core_dims=[["wavelength"], ["wavelength", "sample", "point"]],
-                output_core_dims=[["sample", "point"]],
-                dask="parallelized",
-                output_dtypes=[float],
-            ) / (spectral_albedo.coords["wavelength"].max().values - spectral_albedo.coords["wavelength"].min().values)
+                self.spectral_albedo.coords["wavelength"],
+            ) / (
+                self.spectral_albedo.coords["wavelength"].max().values
+                - self.spectral_albedo.coords["wavelength"].min().values
+            )
             broadband_albedo = xr.DataArray(broadband_albedo, dims=["sample", "point"], name="broadband albedo")
             return broadband_albedo
         else: