+++ Changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Unrelased
+
+### Added
+- added hidden property _capture_spectral which is the orginal representation of the capture data from the spectral library
+
+### Fixed
+
+- an issue when black and white references where mixed up
+
 ## 0.2.0
 
 ### Added